films-for-all/backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.6
django-environ==0.9.0
numpy==1.24.2
pandas==2.1.1
scikit-learn==1.3.0
 